+++ crates/services/upgradable-executor/wasm-executor/src/tx_source.rs
@@ -16,7 +16,7 @@ impl TransactionsSource for WasmTxSource {
     fn next(
         &self,
         gas_limit: u64,
-        tx_count_limit: u16,
+        tx_count_limit: u32,
         block_transaction_size_limit: u32,
     ) -> Vec<MaybeCheckedTransaction> {
         ext::next_transactions(gas_limit, tx_count_limit, block_transaction_size_limit)
